requirements.txt
flask==2.3.3
requests==2.31.0
transformers==4.40.1
torch==2.2.2
python-dotenv==1.0.1